[jit] split TestDataParallel, TestAsync
