+++ test/jit_utils.py
@@ -35,6 +35,9 @@ import textwrap
 
 IN_TRANSITION_TO_PROFILING_GRAPH_EXECUTOR = False
 
+RUN_CUDA = torch.cuda.is_available()
+RUN_CUDA_MULTI_GPU = RUN_CUDA and torch.cuda.device_count() > 1
+
 class ProfilingMode(Enum):
     OFF = 1
     EXECUTOR = 2
@@ -454,7 +457,7 @@ class JitTestCase(TestCase):
             recording_inputs = reference_tensors
 
         # `check_trace` is set to False because check_trace is run with @no_grad
-        # Also, `checkTrace` already does all the checks 
+        # Also, `checkTrace` already does all the checks
         # against python function
         ge = torch.jit.trace(func, input_tensors, check_tolerance=check_tolerance,
                              _force_outplace=_force_outplace, check_trace=False)
